Add files via upload
Contained within the "FitLife_GymWorkouts.ipynb" file is a robust code that encompasses tasks such as data cleaning, preprocessing, exploratory data analysis, and the creation of a recommender system. This system suggests workouts based on critical features like Type, BodyPart, Level, and Rating.

In parallel, the Alteryx file serves as a structured workflow, detailing the steps undertaken in the data understanding and cleaning process. This file offers a clear and comprehensive visualization of the procedures involved in maintaining data integrity and reliability.